add .editorconfig
See https://editorconfig.org/

An .editorconfig file allows text editors and code hosting platforms
to understand how to properly modify and display files.

The primary goal of this patch is to make GitHub display the lines with
an indentation of 4 units.

The meson.build file has been modified to use tabs instead of spaces.
This is to align it with the .editorconfig.

There are also other meson.build files in the libft and libftprintf
subprojects, but I feel like going through the effort of changing these
files as well is not worth it because these files are edited very
infrequently.