input: make struct s_input_formatted_string 'self-contained'
By applying const to the char * in struct s_input_formatted_string we
communicate that and memory management of .string is not the
responsibility of the holder of the struct. Calling free(3) on .string
will now result in a compiler error. This means that the struct can now
be safely stored on the stack without the risk of freeing .string by
accident.