input: fix pasting
Previously the input__read_keypress function read 4 bytes at once,
assuming that a regular character would only be in the first byte
because a user couldn't type so fast. This assumption was flawed as
pasting text would cause all 4 bytes to be filled with data. The result
would be that, if pasting text, some but not all characters would
actually be recorded.

This patch switches to a new system which reads the input byte by byte
instead. This will also allow parsing of more complex control codes in
the future.