input: fix graphical glitch when terminal scrolls up
The constant restoring to the saved position every keypress doesn't
work. The issue is that when the terminal scrolls up the saved position
isn't moved up as well. To solve this we would need to move up the saved
position every time we scroll up. With the current approach that would
require us to read out the cursor position on every keypress to check
if we there is enough space and therefore would have moved up.

Unfortunately that requires sending out an escape sequence and reading
the response. Not only is this very slow and this causes flickering,
this is also unreliable because the user may enter characters before
the escape sequence is send back.

We can't just switch to keeping track of our position ourselves because
saving the cursor position was the most reliable way we found of
restoring the cursor position after a terminal resize.

This new approach will use both techniques. We will keep track of the
cursor position ourselves when redrawing the prompt in response to a
keypress. If we get the SIGWINCH (window size change) signal we will
restore the cursor position. If we calculate at that we scrolled up
during the printing of user input, we will move the saved position up.

This new technique should work in most situations. Some cases are still
unaccounted for, like what happens if the the terminal scrolls up
because of a terminal resize, but it should be enough for 21sh.