tosh: treat input event loop errors as non-fatal
If the user types at the same time as resizing the window the code which
reads out the current cursor position gets confused and spits out an
error. Right now this causes tosh to exit - make this error non-fatal.

Steps to reproduce:

On macOS:

 1. In iTerm2, click on Window > Zoom
 2. Start spamming random text while the zoom is in progress

On Linux:

 1. Launch two terminals
 2. Run "while true; do kill -WINCH $(pidof tosh) 2>&-; done" in one of
    them
 3. Run tosh in the other
 4. Start to type

Old result:

    $ ./tosh
    tosh: fatal: unable to read input: tosh: failed to run next action: unable to get cursor position: expected terminal response to start with \033[
    $ 

New result:

    $ ./tosh
    tosh: failed to run next action: unable to get cursor position: expected terminal response to start with \033[
    TOSH $